@@ -1,6 +1,6 @@
 [project]
 name = "vector-inspector"
-version = "0.4.1"
+version = "0.4.2"
 description = "A comprehensive desktop application for visualizing, querying, and managing vector database data"
 authors = [
     {name = "Anthony Dawson", email = "anthonypdawson+github@gmail.com"},
@@ -62,11 +62,12 @@ dev = [
     "pytest-cov>=7.0.0",
     "pytest-xdist>=3.8.0",
     "pytest-testmon>=2.2.0",
+    "pytest-timeout>=2.4.0",
 ]
 
 [tool.ruff]
-line-length = 100
-target-version = "py311"
+line-length = 120
+target-version = "py312"
 extend-exclude = ["build", "dist", ".venv"]
 
 [tool.ruff.format]
@@ -116,7 +117,7 @@ env = { PYTHONPATH = "src" }
 [tool.briefcase]
 project_name = "Vector Inspector"
 bundle = "com.divinedevops.vector-inspector"
-version = "0.4.1"
+version = "0.4.2"
 url = "https://vector-inspector.divinedevops.com"
 license.file = "LICENSE"
 author = "Anthony Dawson"
