@@ -16,6 +16,8 @@ from vector_inspector.core.connection_manager import ConnectionInstance, Connect
 from vector_inspector.core.logging import log_error
 from vector_inspector.services.profile_service import ProfileService
 from vector_inspector.services.settings_service import SettingsService
+from vector_inspector.services.task_runner import ThreadedTaskRunner
+from vector_inspector.state import AppState
 from vector_inspector.ui.components.connection_manager_panel import ConnectionManagerPanel
 from vector_inspector.ui.components.profile_manager_panel import ProfileManagerPanel
 from vector_inspector.ui.controllers.connection_controller import ConnectionController
@@ -41,15 +43,17 @@ class MainWindow(InspectorShell):
     def __init__(self):
         super().__init__()
 
+        # Shared application state and task runner
+        self.app_state = AppState()
+        self.task_runner = ThreadedTaskRunner()
+
         # Core services
         self.connection_manager = ConnectionManager()
         self.profile_service = ProfileService()
         self.settings_service = SettingsService()
 
         # Controller for connection operations
-        self.connection_controller = ConnectionController(
-            self.connection_manager, self.profile_service, self
-        )
+        self.connection_controller = ConnectionController(self.connection_manager, self.profile_service, self)
 
         # State
         self.visualization_view = None
@@ -124,7 +128,7 @@ class MainWindow(InspectorShell):
         tab_defs = InspectorTabs.get_standard_tabs()
 
         for i, tab_def in enumerate(tab_defs):
-            widget = InspectorTabs.create_tab_widget(tab_def, connection=None)
+            widget = InspectorTabs.create_tab_widget(tab_def, app_state=self.app_state, task_runner=self.task_runner)
             self.add_main_tab(widget, tab_def.title)
 
             # Store references to views (except placeholder)
@@ -263,9 +267,7 @@ class MainWindow(InspectorShell):
         # Update indicator label (hidden by default)
         self.update_indicator = QLabel()
         self.update_indicator.setText("")
-        self.update_indicator.setStyleSheet(
-            "color: #2980b9; font-weight: bold; text-decoration: underline;"
-        )
+        self.update_indicator.setStyleSheet("color: #2980b9; font-weight: bold; text-decoration: underline;")
         self.update_indicator.setVisible(False)
         self.update_indicator.setCursor(Qt.PointingHandCursor)
         self.statusBar().addPermanentWidget(self.update_indicator)
@@ -288,9 +290,7 @@ class MainWindow(InspectorShell):
             if latest:
                 current_version = get_app_version()
                 latest_version = latest.get("tag_name")
-                if latest_version and UpdateService.compare_versions(
-                    current_version, latest_version
-                ):
+                if latest_version and UpdateService.compare_versions(current_version, latest_version):
 
                     def show_update():
                         self._latest_release = latest
@@ -388,12 +388,8 @@ class MainWindow(InspectorShell):
     def _connect_signals(self):
         """Connect signals between components."""
         # Connection manager signals
-        self.connection_manager.active_connection_changed.connect(
-            self._on_active_connection_changed
-        )
-        self.connection_manager.active_collection_changed.connect(
-            self._on_active_collection_changed
-        )
+        self.connection_manager.active_connection_changed.connect(self._on_active_connection_changed)
+        self.connection_manager.active_collection_changed.connect(self._on_active_collection_changed)
         self.connection_manager.collections_updated.connect(self._on_collections_updated)
         self.connection_manager.connection_opened.connect(self._on_connection_opened)
 
@@ -407,16 +403,12 @@ class MainWindow(InspectorShell):
         # Profile panel signals
         self.profile_panel.connect_profile.connect(self._connect_to_profile)
 
-    def _on_connection_completed(
-        self, connection_id: str, success: bool, collections: list, error: str
-    ):
+    def _on_connection_completed(self, connection_id: str, success: bool, collections: list, error: str):
         """Handle connection completed event from controller."""
         if success:
             # Switch to Active connections tab
             self.set_left_panel_active(0)
-            self.statusBar().showMessage(
-                f"Connected successfully ({len(collections)} collections)", 5000
-            )
+            self.statusBar().showMessage(f"Connected successfully ({len(collections)} collections)", 5000)
 
     def _on_tab_changed(self, index: int):
         """Handle tab change - lazy load visualization tab."""
@@ -424,26 +416,20 @@ class MainWindow(InspectorShell):
             # Lazy load visualization view
             from vector_inspector.ui.views.visualization_view import VisualizationView
 
-            # Get active connection
-            active = self.connection_manager.get_active_connection()
-
-            self.visualization_view = VisualizationView(active)
+            self.visualization_view = VisualizationView(self.app_state, self.task_runner)
 
             # Connect signal to view point in data browser
-            self.visualization_view.view_in_data_browser_requested.connect(
-                self._on_view_in_data_browser_requested
-            )
+            self.visualization_view.view_in_data_browser_requested.connect(self._on_view_in_data_browser_requested)
 
             # Replace placeholder with actual view
             self.remove_main_tab(InspectorTabs.VISUALIZATION_TAB)
-            self.add_main_tab(
-                self.visualization_view, "Visualization", InspectorTabs.VISUALIZATION_TAB
-            )
+            self.add_main_tab(self.visualization_view, "Visualization", InspectorTabs.VISUALIZATION_TAB)
             self.set_main_tab_active(InspectorTabs.VISUALIZATION_TAB)
 
-            # Set collection if one is already selected
-            if active and active.active_collection:
-                self.visualization_view.set_collection(active.active_collection)
+            # Set collection if one is already selected (for initial state)
+            # Future collection changes will be handled by app_state.collection_changed signal
+            if self.app_state.collection:
+                self.visualization_view.set_collection(self.app_state.collection)
 
     def _on_active_connection_changed(self, connection_id):
         """Handle active connection change."""
@@ -503,17 +489,25 @@ class MainWindow(InspectorShell):
 
     def _update_views_with_connection(self, connection: Optional[ConnectionInstance]):
         """Update all views with a new connection."""
-        # Clear current collection when switching connections
+        # Update AppState (new pattern - triggers reactive views)
+        # AppState exposes properties rather than setter methods.
+        self.app_state.provider = connection
+
+        # Clear current collection when switching connections (legacy pattern)
         self.info_panel.current_collection = None
-        self.metadata_view.current_collection = None
-        self.search_view.current_collection = None
+        if hasattr(self.metadata_view, "current_collection"):
+            self.metadata_view.current_collection = None
+        if hasattr(self.search_view, "current_collection"):
+            self.search_view.current_collection = None
         if self.visualization_view is not None:
             self.visualization_view.current_collection = None
 
-        # Update connection references
+        # Update connection references (legacy pattern)
         self.info_panel.connection = connection
-        self.metadata_view.connection = connection
-        self.search_view.connection = connection
+        if hasattr(self.metadata_view, "connection"):
+            self.metadata_view.connection = connection
+        if hasattr(self.search_view, "connection"):
+            self.search_view.connection = connection
 
         if self.visualization_view is not None:
             self.visualization_view.connection = connection
@@ -529,9 +523,17 @@ class MainWindow(InspectorShell):
             active = self.connection_manager.get_active_connection()
             database_name = active.id if active else ""
 
+            # Update AppState (new pattern - triggers reactive views)
+            # AppState exposes properties rather than setter methods.
+            self.app_state.collection = collection_name
+            self.app_state.database = database_name
+
+            # Update views (legacy pattern - for views not yet refactored)
             self.info_panel.set_collection(collection_name, database_name)
-            self.metadata_view.set_collection(collection_name, database_name)
-            self.search_view.set_collection(collection_name, database_name)
+            if hasattr(self.metadata_view, "set_collection"):
+                self.metadata_view.set_collection(collection_name, database_name)
+            if hasattr(self.search_view, "set_collection"):
+                self.search_view.set_collection(collection_name, database_name)
 
             if self.visualization_view is not None:
                 self.visualization_view.set_collection(collection_name)
@@ -591,6 +593,11 @@ class MainWindow(InspectorShell):
     def _toggle_cache(self, checked: bool):
         """Toggle caching on/off."""
         self.settings_service.set_cache_enabled(checked)
+        # Update cache manager state (AppState coordination)
+        if checked:
+            self.app_state.cache_manager.enable()
+        else:
+            self.app_state.cache_manager.disable()
         status = "enabled" if checked else "disabled"
         self.statusBar().showMessage(f"Caching {status}", 3000)
 
@@ -616,9 +623,7 @@ class MainWindow(InspectorShell):
         # Get active connection
         active = self.connection_manager.get_active_connection()
         if not active or not active.is_connected:
-            QMessageBox.information(
-                self, "No Connection", "Please connect to a database first to create a collection."
-            )
+            QMessageBox.information(self, "No Connection", "Please connect to a database first to create a collection.")
             return
 
         # Show dialog and create collection
